requirements.txt
langgraph
pydantic
openai
python-dotenv 